+++ satellite/src/main/java/com/linbit/linstor/layer/drbd/DrbdLayer.java
@@ -16,6 +16,7 @@ import com.linbit.extproc.ExtCmd.OutputData;
 import com.linbit.extproc.ExtCmdFactory;
 import com.linbit.extproc.ExtCmdFailedException;
 import com.linbit.linstor.InternalApiConsts;
+import com.linbit.linstor.PriorityProps;
 import com.linbit.linstor.annotation.DeviceManagerContext;
 import com.linbit.linstor.api.ApiCallRcImpl;
 import com.linbit.linstor.api.ApiConsts;
@@ -33,6 +34,7 @@ import com.linbit.linstor.core.objects.Resource;
 import com.linbit.linstor.core.objects.Resource.Flags;
 import com.linbit.linstor.core.objects.ResourceDefinition;
 import com.linbit.linstor.core.objects.Snapshot;
+import com.linbit.linstor.core.objects.StorPool;
 import com.linbit.linstor.core.objects.Volume;
 import com.linbit.linstor.core.pojos.LocalPropsChangePojo;
 import com.linbit.linstor.dbdrivers.DatabaseException;
@@ -68,6 +70,7 @@ import com.linbit.linstor.storage.kinds.DeviceProviderKind;
 import com.linbit.linstor.storage.utils.MkfsUtils;
 import com.linbit.linstor.storage.utils.VolumeUtils;
 import com.linbit.linstor.utils.layer.DrbdLayerUtils;
+import com.linbit.linstor.utils.layer.LayerVlmUtils;
 import com.linbit.utils.AccessUtils;
 
 import javax.inject.Inject;
@@ -217,7 +220,7 @@ public class DrbdLayer implements DeviceLayer
             StateFlags<Flags> rscFlags = drbdVlmData.getRscLayerObject().getAbsResource().getStateFlags();
             boolean isDiskless = rscFlags.isSet(workerCtx, Resource.Flags.DRBD_DISKLESS) &&
                 !rscFlags.isSomeSet(workerCtx, Resource.Flags.DISK_ADD_REQUESTED, Resource.Flags.DISK_ADDING);
-            if (!isDiskless)
+            if (!isDiskless && !isSkipDiskEnabled(drbdRscData))
             {
                 long netSize = drbdVlmData.getUsableSize();
 
@@ -308,7 +311,7 @@ public class DrbdLayer implements DeviceLayer
         {
             boolean isDiskless = drbdVlmData.getRscLayerObject().getAbsResource().getStateFlags()
                 .isSet(workerCtx, Resource.Flags.DRBD_DISKLESS);
-            if (!isDiskless)
+            if (!isDiskless && !isSkipDiskEnabled(drbdRscData))
             {
                 // let next layer calculate
                 VlmProviderObject<Resource> dataChildVlmData = drbdVlmData.getChildBySuffix(
@@ -632,7 +635,9 @@ public class DrbdLayer implements DeviceLayer
 
             shrinkVolumesIfNecessary(drbdRscData);
 
-            if (!childAlreadyProcessed)
+            final boolean skipDisk = isSkipDiskEnabled(drbdRscData);
+
+            if (!childAlreadyProcessed && !skipDisk)
             {
                 contProcess = processChild(drbdRscData, snapshotList, apiCallRc);
             }
@@ -670,9 +675,9 @@ public class DrbdLayer implements DeviceLayer
             {
                 // hasMetaData needs to be run after child-resource processed
                 List<DrbdVlmData<Resource>> createMetaData = new ArrayList<>();
-                if (!drbdRscData.getAbsResource().isDrbdDiskless(workerCtx))
+                if (!drbdRscData.getAbsResource().isDrbdDiskless(workerCtx) && !skipDisk)
                 {
-                    // do not try to create meta data while the resource is diskless....
+                    // do not try to create meta data while the resource is diskless or skipDisk is enabled
                     for (DrbdVlmData<Resource> drbdVlmData : checkMetaData)
                     {
                         if (
@@ -808,7 +813,7 @@ public class DrbdLayer implements DeviceLayer
                     drbdUtils.adjust(
                         drbdRscData,
                         false,
-                        false,
+                        skipDisk,
                         false
                     );
 
@@ -858,7 +863,7 @@ public class DrbdLayer implements DeviceLayer
                         drbdUtils.adjust(
                             drbdRscData,
                             false,
-                            false,
+                            skipDisk,
                             false
                         );
                     }
@@ -1960,5 +1965,23 @@ public class DrbdLayer implements DeviceLayer
     {
         // ignored
         return null;
+    }
+
+    private boolean isSkipDiskEnabled(DrbdRscData<Resource> drbdRscData) throws AccessDeniedException
+    {
+        Resource rsc = drbdRscData.getAbsResource();
+        PriorityProps prioProps = new PriorityProps(rsc.getProps(workerCtx));
+        for (StorPool storPool : LayerVlmUtils.getStorPools(rsc, workerCtx))
+        {
+            prioProps.addProps(storPool.getProps(workerCtx));
+        }
+        prioProps.addProps(
+            rsc.getNode().getProps(workerCtx),
+            rsc.getResourceDefinition().getProps(workerCtx),
+            stltCfgAccessor.getReadonlyProps()
+        );
+
+        return ApiConsts.VAL_TRUE
+            .equals(prioProps.getProp(ApiConsts.KEY_DRBD_SKIP_DISK, ApiConsts.NAMESPC_DRBD_OPTIONS));
     }
 }