+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed unrecoverable state with failed rollback of deleting snapshot
 - Fixed race condition between node lost and backup queue cleanup
 - NetCom SSL reimplementation: Fixed buffer handling, SSL handshake/negotiation/renegotiation
+- Prohibit disk accesses in DrbdLayer while SkipDisk is set
 
 ### Removed